refactor(mobile): remove image processing from the post edit step
