+++ apps/ledger-live-mobile/src/screens/CustomImage/PreviewPostEdit.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useRef, useState } from "react";
+import React, { useCallback } from "react";
 import { Button, Flex, InfiniteLoader } from "@ledgerhq/native-ui";
 import { ImagePreviewError } from "@ledgerhq/live-common/customImage/errors";
 import { NativeSyntheticEvent, ImageErrorEventData } from "react-native";
@@ -12,11 +12,6 @@ import {
 import { CustomImageNavigatorParamList } from "../../components/RootNavigator/types/CustomImageNavigator";
 import { NavigatorName, ScreenName } from "../../const";
 import FramedImage from "../../components/CustomImage/FramedImage";
-import ImageProcessor, {
-  Props as ImageProcessorProps,
-  ProcessorPreviewResult,
-  ProcessorRawResult,
-} from "../../components/CustomImage/ImageProcessor";
 
 type NavigationProps = BaseComposite<
   StackNavigatorProps<
@@ -28,7 +23,7 @@ type NavigationProps = BaseComposite<
 const PreviewPostEdit = ({ navigation, route }: NavigationProps) => {
   const { t } = useTranslation();
   const { params } = route;
-  const { baseImageFile, contrast, imagePreview, device } = params;
+  const { baseImageFile, imagePreview, imageData, device } = params;
 
   const handleError = useCallback(
     (error: Error) => {
@@ -38,39 +33,20 @@ const PreviewPostEdit = ({ navigation, route }: NavigationProps) => {
     [navigation, device],
   );
 
-  const [previewLoading, setPreviewLoading] = useState<boolean>(true);
-  const [processorPreviewImage, setProcessorPreviewImage] =
-    useState<ProcessorPreviewResult | null>(null);
-  const [rawResultLoading, setRawResultLoading] = useState(false);
-  const imageProcessorRef = useRef<ImageProcessor>(null);
-
-  const handlePreviewResult: ImageProcessorProps["onPreviewResult"] =
-    useCallback(
-      data => {
-        setProcessorPreviewImage(data);
-        setPreviewLoading(false);
-      },
-      [setProcessorPreviewImage],
-    );
-
-  const handleRawResult: ImageProcessorProps["onRawResult"] = useCallback(
-    (data: ProcessorRawResult) => {
-      if (!processorPreviewImage) {
-        /**
-         * this should not happen as the "request raw result" button is only
-         * visible once the preview is there
-         * */
-        throw new ImagePreviewError();
-      }
-      navigation.navigate(ScreenName.CustomImageStep3Transfer, {
-        rawData: data,
-        previewData: processorPreviewImage,
-        device,
-      });
-      setRawResultLoading(false);
-    },
-    [navigation, setRawResultLoading, processorPreviewImage, device],
-  );
+  const handleSetPicture = useCallback(() => {
+    if (!imagePreview) {
+      /**
+       * this should not happen as the "request raw result" button is only
+       * visible once the preview is there
+       * */
+      throw new ImagePreviewError();
+    }
+    navigation.navigate(ScreenName.CustomImageStep3Transfer, {
+      rawData: imageData,
+      previewData: imagePreview,
+      device,
+    });
+  }, [navigation, device, imagePreview, imageData]);
 
   const handlePreviewImageError = useCallback(
     ({ nativeEvent }: NativeSyntheticEvent<ImageErrorEventData>) => {
@@ -79,11 +55,6 @@ const PreviewPostEdit = ({ navigation, route }: NavigationProps) => {
     },
     [handleError],
   );
-
-  const requestRawResult = useCallback(() => {
-    imageProcessorRef?.current?.requestRawResult();
-    setRawResultLoading(true);
-  }, [imageProcessorRef, setRawResultLoading]);
 
   const handleEditPicture = useCallback(() => {
     navigation.navigate(NavigatorName.CustomImage, {
@@ -104,16 +75,6 @@ const PreviewPostEdit = ({ navigation, route }: NavigationProps) => {
     <SafeAreaView style={{ flex: 1 }} edges={["bottom"]}>
       <Flex flex={1}>
         <Flex flex={1}>
-          {imagePreview?.imageBase64DataUri && (
-            <ImageProcessor
-              ref={imageProcessorRef}
-              imageBase64DataUri={imagePreview?.imageBase64DataUri}
-              onPreviewResult={handlePreviewResult}
-              onError={handleError}
-              onRawResult={handleRawResult}
-              contrast={contrast}
-            />
-          )}
           <Flex
             flex={1}
             flexDirection="column"
@@ -123,20 +84,12 @@ const PreviewPostEdit = ({ navigation, route }: NavigationProps) => {
             <FramedImage
               onError={handlePreviewImageError}
               fadeDuration={0}
-              source={{ uri: processorPreviewImage?.imageBase64DataUri }}
+              source={{ uri: imagePreview?.imageBase64DataUri }}
             />
           </Flex>
         </Flex>
         <Flex px={8}>
-          <Button
-            type="main"
-            size="large"
-            mb={4}
-            disabled={previewLoading}
-            pending={rawResultLoading}
-            displayContentWhenPending
-            onPress={requestRawResult}
-          >
+          <Button type="main" size="large" mb={4} onPress={handleSetPicture}>
             {t("customImage.preview.setPicture")}
           </Button>
           <Button mb={8} size="large" onPress={handleEditPicture}>
